requirements.txt
fastapi==0.110.0
uvicorn==0.22.0
pydantic==1.10.13
starlette==0.36.3
typing-extensions==4.15.0

sentence-transformers==2.2.2
transformers==4.40.0
huggingface-hub==0.36.0

faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.12.0
nltk==3.9.1

torch==2.2.0
torchvision==0.17.0

python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.3
regex==2023.12.25

# pip install sentence-transformers==2.2.2 huggingface-hub==0.14.1
